third_party/blink/web_tests/external/wpt/fetch/http-cache/heuristic-expected.txt

This is a testharness.js-based test.
FAIL HTTP cache reuses an unknown response with Last-Modified based upon heuristic freshness when Cache-Control: public is present assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
PASS HTTP cache does not reuse an unknown response with Last-Modified based upon heuristic freshness when Cache-Control: public is not present
PASS HTTP cache reuses a 200 OK response with Last-Modified based upon heuristic freshness
PASS HTTP cache reuses a 203 Non-Authoritative Information response with Last-Modified based upon heuristic freshness
FAIL HTTP cache reuses a 204 No Content response with Last-Modified based upon heuristic freshness assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
FAIL HTTP cache reuses a 404 Not Found response with Last-Modified based upon heuristic freshness assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
FAIL HTTP cache reuses a 405 Method Not Allowed response with Last-Modified based upon heuristic freshness assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
PASS HTTP cache reuses a 410 Gone response with Last-Modified based upon heuristic freshness
FAIL HTTP cache reuses a 414 URI Too Long response with Last-Modified based upon heuristic freshness assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
FAIL HTTP cache reuses a 501 Not Implemented response with Last-Modified based upon heuristic freshness assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
PASS HTTP cache does not use a 201 Created response with Last-Modified based upon heuristic freshness
PASS HTTP cache does not use a 202 Accepted response with Last-Modified based upon heuristic freshness
PASS HTTP cache does not use a 403 Forbidden response with Last-Modified based upon heuristic freshness
PASS HTTP cache does not use a 502 Bad Gateway response with Last-Modified based upon heuristic freshness
PASS HTTP cache does not use a 503 Service Unavailable response with Last-Modified based upon heuristic freshness
PASS HTTP cache does not use a 504 Gateway Timeout response with Last-Modified based upon heuristic freshness
Harness: the test ran to completion.
